fix: Process schema fields in custom resource actions for CKAN 2.11 compatibility
Move schema field processing (schema_json, schema_url, schema_upload) from
plugin hooks to custom resource_create/resource_update actions to ensure they
work correctly when resources are created/updated via web forms in CKAN 2.11.

Changes:
- Extract process_schema_fields() to utils.py for reuse
- Add schema processing to resource_create and resource_update actions
- Remove non-existent before_dataset_update hook from plugin
- Update plugin hooks to use centralized utility function

This fixes 6 form tests that were failing with KeyError: 'schema' because
the before_dataset_update hook doesn't exist in CKAN 2.11's IPackageController.
